services/worker/requirements.txt
mysql-connector-python==9.0.0
redis==5.0.8